requirements.txt
PySide6>=6.7
pypdfium2>=4.30
Pillow>=10.3
pywin32>=306
